Copy Read Me.rtf into Perian.component
That way the license and link to the website are kept with the component,
which helps people avoid accidentally violating the license.

NOTE: All such people are still violating LGPL section 4, which requires
distributing source from the same place as the binary. It does not allow
linking back to perian.org.


git-svn-id: http://svn.perian.org/trunk@1309 621663c8-3916-0410-8f58-edc14a8543d5