Prevent false FormatInfo failures and align error handling in audio import better
FFission:
- Don't fail on FormatInfo requests that aren't 44100khz/2ch (oops)
This fix is slightly problematic as the ASBD created is impossible, but we can
watch out for problems with that as only our code is going to see it.

Audio importers:
- Don't fail if FormatInfo fails. The audio will probably be unplayable
but at least we can look at the output in Atom Inspector.
- Do fail if channels/codec ID are set, because we'll probably crash later
if so.
Slightly hypocritical...

Fixes #552
Fixes MKV+48khz Vorbis import



git-svn-id: http://svn.perian.org/trunk@1389 621663c8-3916-0410-8f58-edc14a8543d5